OSX 에서 컴파일할 때에 warning / error 가 발생하지 않도록 수정함.

+++ SipStack/SipStackSetup.cpp
@@ -82,7 +82,7 @@ bool CSipStackSetup::Check( )
 	
 	gettimeofday( &sttTime, NULL );
 
-	snprintf( szTemp, sizeof(szTemp), "%s:%d:%d:%d", m_strLocalIp.c_str(), m_iLocalUdpPort, sttTime.tv_sec, sttTime.tv_usec );
+	snprintf( szTemp, sizeof(szTemp), "%s:%d:%d:%d", m_strLocalIp.c_str(), m_iLocalUdpPort, (int)sttTime.tv_sec, sttTime.tv_usec );
 	SipMd5String16( szTemp, szMd5 );
 
 	SipSetSystemId( szMd5 );
